User Interface and editing of style lines in array
The style-swapper-view is mostly copied from the command-palette
package and works well. It is reduced since not all functions are
needed (e.g.  persistence).

The Stylesheet gets read and options displayed.

Lines get edited in local line-array. Need to write this array back
to the style sheet.

Need to clean up the if conditions in main loop, as well as implement
all commented todo notes.